Create bootloader to kernel stage functions
